[2025.05.13] - 김건우 1. 매매일지 매도 로직 수정

+++ src/main/java/com/kwfw/findiary/model/DiaryDto.java
@@ -8,7 +8,7 @@ public class DiaryDto {
     private int trading_count;
     private int status;
 
-    private long trading_price;
+    private float trading_price;
 
     private String ticker;
     private String stock_name;
